Create Sol4.java
4. Given the head of a linked list which stores only binary values i.e. 1’s and 0’s. In the linked list make sure
that between every 2 nodes containing value 1, there exist an even number of 0’s. If the number of 0’s is odd
then insert a node with value 0 in between the two 1’s. Assume that the first and the last node contains value
The first line of input contains the size of the linked list.The second line of input contains the elements of the
linked list.
INPUT
6
1 0 0 1 0 1
UTPUT
1->0->0->1->0->0->1->NULJ